Update Read me
Hey i need a bit of help. So I have everything set up and I play with one of my friends and for some reason when I knock him into the void it kills him and puts him back in world spawn and doesn’t register the kill so it still says on the scoreboard 2 players left and the only way to win is for me or my friend to do /sw quit. And theres no error messages im playing on 1.19 if that matters